+++ src/native_functions_PAL.h
@@ -1,0 +1,11 @@
+/**
+ * @file native_functions_PAL.h
+ * @brief Pointers to in game functions (PAL)
+ */
+
+#ifndef MML_NATIVE_FUNCTIONS_PAL_H
+#define MML_NATIVE_FUNCTIONS_PAL_H
+
+#error native_functions.h not compatible with PAL
+
+#endif